requirements.txt
apischema==0.15.6
APScheduler==3.6.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.11
idna==3.3
kinopoisk-api-unofficial-client==0.9.8
pkg-resources==0.0.0
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
